FIx: Cambio auth de manager a app
